requirements-dev.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.2.2
gunicorn==23.0.0
idna==3.10
PyJWT==2.4.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
psycopg2-binary==2.9.9
python-dotenv
virtualenv

pytest
pytest-cov
pytest-django
model_bakery
pytest-json-report